fix LayoutNavigation file position
